Develop

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -25,7 +25,7 @@ class CreateUsersTable extends Migration
             $table->string('birth_place')->nullable();
             $table->string('hobbit')->nullable();
             $table->string('introduce')->nullable();
-            $table->integer('is_active')->default(0);
+            $table->tinyInteger('is_active')->default(0);
             $table->integer('role_id')->index()->unsigned()->nullable();
             $table->rememberToken();
             $table->timestamps();
